+++ CTP/shell/src/com/navercorp/cubridqa/shell/deploy/TestCaseSVN.java
@@ -72,7 +72,9 @@ public class TestCaseSVN {
 			String toStar = CommonUtils.concatFile(wsRoot, "*").replace('\\', '/');
 
 			scripts.addCommand("mkdir -p " + wsRoot);
-			scripts.addCommand("rm -rf " + toStar);
+			if (!CommonUtils.isEmpty(toStar)) {
+				scripts.addCommand("rm -rf " + toStar);
+			}
 			scripts.addCommand("cp -r " + fromStar + " " + wsRoot);
 			
 			String result;
